+++ src/main/java/com/kodar/academy/Library/controller/UserController.java
@@ -36,7 +36,7 @@ public class UserController {
     @GetMapping("/users/{id}")
     public ResponseEntity<UserResponseDTO> getUserById(@PathVariable("id") int id) {
         UserResponseDTO userResponseDTO = userService.getUserById(id);
-        if(userResponseDTO.getUsername() == null) {
+        if(userResponseDTO == null) {
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
         return new ResponseEntity<>(userResponseDTO, HttpStatus.OK);
@@ -44,18 +44,27 @@ public class UserController {
 
     @DeleteMapping("/users/delete/{id}")
     public ResponseEntity<String> deleteUserById(@PathVariable("id") int id) {
+        if(userService.getUserById(id) == null) {
+            return new ResponseEntity<>("User with that id doesn't exist", HttpStatus.NOT_FOUND);
+        }
         userService.deleteUser(id);
         return new ResponseEntity<>("User successfully deleted", HttpStatus.OK);
     }
 
     @PutMapping("/users/edit/{id}")
-    public ResponseEntity<UserResponseDTO> editUser(@PathVariable("id") int id, @Valid @RequestBody UserEditDTO userEditDTO) {
+    public ResponseEntity<?> editUser(@PathVariable("id") int id, @Valid @RequestBody UserEditDTO userEditDTO) {
+        if(userService.getUserById(id) == null) {
+            return new ResponseEntity<>("User with that id doesn't exist", HttpStatus.NOT_FOUND);
+        }
         UserResponseDTO userResponseDTO = userService.updateUser(id, userEditDTO);
         return new ResponseEntity<>(userResponseDTO, HttpStatus.OK);
     }
 
     @PutMapping("/users/change-password/{id}")
     public ResponseEntity<String> changePassword(@PathVariable("id") int id, @RequestBody UserCPDTO userCPDTO) {
+        if(userService.getUserById(id) == null) {
+            return new ResponseEntity<>("User with that id doesn't exist", HttpStatus.NOT_FOUND);
+        }
         userService.changePassword(id, userCPDTO);
         return new ResponseEntity<>("Password changed successfully", HttpStatus.OK);
     }
